# Conflicts:
#	src/afyber/shmupfeaturecreep/engine/world/WorldMiddleman.java
#	src/afyber/shmupfeaturecreep/game/EnemyShipParent.java

+++ src/afyber/shmupfeaturecreep/engine/Screen.java
@@ -97,16 +97,16 @@ public class Screen {
 		}
 	}
 
-	public static void draw(String spriteName, float x, float y, int depth) {
+	public static void draw(String spriteName, double x, double y, int depth) {
 		draw(spriteName, x, y, 1, 1, depth);
 	}
-	public static void draw(String spriteName, float x, float y, float xScale, float yScale, int depth) {
+	public static void draw(String spriteName, double x, double y, double xScale, double yScale, int depth) {
 		draw(spriteName, x, y, xScale, yScale, depth, 1);
 	}
-	public static void draw(String spriteName, float x, float y, float xScale, float yScale, int depth, float alpha) {
+	public static void draw(String spriteName, double x, double y, double xScale, double yScale, int depth, double alpha) {
 		if (isDrawing) {
 			try {
-				drawRequests.add(new DrawRequest(spriteName, Math.round(x), Math.round(y), xScale, yScale, depth, alpha));
+				drawRequests.add(new DrawRequest(spriteName, (int)Math.round(x), (int)Math.round(y), xScale, yScale, depth, alpha));
 			} catch (NullPointerException e) {
 				MainClass.LOGGER.log(LoggingLevel.ERROR, "Draw attempted before 'drawRequests' initialized", e);
 			}
@@ -161,10 +161,10 @@ public class Screen {
 
 	private static void copySpriteRegionToFrame(SpriteSheetRegion sprite, DrawRequest request) {
 		// all this to say, if the sprite does not overlap the image, do not draw it
-		int actualX = request.x() - Math.round(sprite.originX() * request.xScale());
-		int actualY = request.y() - Math.round(sprite.originY() * request.yScale());
-		int actualX2 = actualX + Math.round(sprite.dataWidth() * request.xScale());
-		int actualY2 = actualY + Math.round(sprite.dataHeight() * request.yScale());
+		int actualX = request.x() - (int)Math.round(sprite.originX() * request.xScale());
+		int actualY = request.y() - (int)Math.round(sprite.originY() * request.yScale());
+		int actualX2 = actualX + (int)Math.round(sprite.dataWidth() * request.xScale());
+		int actualY2 = actualY + (int)Math.round(sprite.dataHeight() * request.yScale());
 		if (!GeneralUtil.areRectanglesIntersecting(actualX, actualY, actualX2, actualY2, 0, 0, MainClass.WINDOW_WIDTH - 1, MainClass.WINDOW_HEIGHT - 1)) {
 			return;
 		}
@@ -173,7 +173,7 @@ public class Screen {
 
 		SpriteSheetRegion scaledSprite = scaleImageData(sprite.data(), request.xScale(), request.yScale(), sprite.originX(), sprite.originY());
 		int[][] spriteData = scaledSprite.data();
-		float alphaPercent = Math.min(1, request.alpha());
+		double alphaPercent = Math.min(1, request.alpha());
 
 		for (int y = 0; y < scaledSprite.dataHeight(); y++) {
 			for (int x = 0; x < scaledSprite.dataWidth(); x++) {
@@ -190,7 +190,7 @@ public class Screen {
 				}
 				else if ((spriteData[y][x] >> 24 & 0xFF) != 0x0 && alphaPercent > 0) {
 					// basically takes a weighted average of the R, G, and B components of the sprite and the current frame, with the weight being the alpha of the sprite being drawn
-					float percentage = ((float)(spriteData[y][x] >> 24 & 0xFF) / 0xFF) * alphaPercent;
+					double percentage = ((float)(spriteData[y][x] >> 24 & 0xFF) / 0xFF) * alphaPercent;
 					currentFrame[calculatedY][calculatedX] = 0xFF000000 | Math.min(0xFF, (int)((currentFrame[calculatedY][calculatedX] >> 16 & 0xFF) * (1 - percentage) + (spriteData[y][x] >> 16 & 0xFF) * percentage)) << 16 | Math.min(0xFF, (int)((currentFrame[calculatedY][calculatedX] >> 8 & 0xFF) * (1 - percentage) + (spriteData[y][x] >> 8 & 0xFF) * percentage)) << 8 | Math.min(0xFF, (int)((currentFrame[calculatedY][calculatedX] & 0xFF) * (1 - percentage) + (spriteData[y][x] & 0xFF) * percentage));
 				}
 			}
@@ -202,9 +202,8 @@ public class Screen {
 		image.setRGB(0, 0, MainClass.WINDOW_WIDTH, MainClass.WINDOW_HEIGHT, singleArray.a(), 0, singleArray.dataWidth());
 	}
 
-	public static SpriteSheetRegion scaleImageData(int[][] data, float xScale, float yScale, int originX, int originY) {
+	public static SpriteSheetRegion scaleImageData(int[][] data, double xScale, double yScale, int originX, int originY) {
 		// TODO: properly figure out scaling to prevent annoying jerkiness when constantly changing scale with an origin not equal to 0,0
-		// TODO: when the new width/height should be exactly equal to an int, it probably won't actually be because of a floating point precision problem
 		if (xScale == 1 && yScale == 1) {
 			return new SpriteSheetRegion(data, data[0].length, data.length, originX, originY);
 		}
@@ -223,7 +222,7 @@ public class Screen {
 
 		int[][] newDataX;
 
-		float runningTally = 0;
+		double runningTally = 0;
 		int intTally;
 		int newXOrigin = 0;
 		int newX = data[0].length;
@@ -302,7 +301,7 @@ public class Screen {
 		return region;
 	}
 
-	public static SpriteSheetRegion getSpriteScaled(String spriteName, float xScale, float yScale) {
+	public static SpriteSheetRegion getSpriteScaled(String spriteName, double xScale, double yScale) {
 		SpriteSheetRegion region = null;
 		if (allSprites.containsKey(spriteName)) {
 			region = allSprites.get(spriteName);
@@ -323,7 +322,7 @@ public class Screen {
 	}
 
 	// TODO: check accuracy against getSpriteScaled
-	public static SpriteInformation getScaledSpriteInfo(String spriteName, float xScale, float yScale) {
+	public static SpriteInformation getScaledSpriteInfo(String spriteName, double xScale, double yScale) {
 		SpriteInformation info = getSpriteInfo(spriteName);
 		if (info != null) {
 			int newOriginX;
@@ -374,5 +373,5 @@ public class Screen {
 		}
 	}
 
-	private record DrawRequest(String spriteName, int x, int y, float xScale, float yScale, int depth, float alpha) {}
+	private record DrawRequest(String spriteName, int x, int y, double xScale, double yScale, int depth, double alpha) {}
 }
