+++ src/afyber/shmupfeaturecreep/engine/rooms/ObjectCreationReference.java
@@ -1,3 +1,3 @@
 package afyber.shmupfeaturecreep.engine.rooms;
 
-public record ObjectCreationReference(String classOfObject, float x, float y, float imageXScale, float imageYScale, int depth) {}
+public record ObjectCreationReference(String classOfObject, double x, double y, double imageXScale, double imageYScale, int depth) {}
